refactor(factory): create a new faculty factory, update imports
