+++ fiesta/apps/utils/management/commands/seed.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import djclick as click
 
-from apps.utils.factories.universities import FacultyFactory
+from apps.utils.factories.faculty import FacultyFactory
 
 
 @click.command()
